requirements.txt
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
elastic-transport==8.15.0
elasticsearch==8.15.1
elasticsearch-dsl==8.15.3
fastapi==0.115.0
h11==0.14.0
idna==3.10
pika==1.3.2
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
